3-Kapselung-des-Gamepads

See merge request r_bernhard/Ipponboard!4

+++ base/FightCategoryManagerDlg.cpp
@@ -109,7 +109,7 @@ void FightCategoryManagerDlg::on_pushButton_add_pressed()
 void FightCategoryManagerDlg::load_values()
 //---------------------------------------------------------
 {
-	for (int i(0); i < m_pClassMgr->CategoryCount(); ++i)
+	for (int i(0); i < (int)m_pClassMgr->CategoryCount(); ++i)
 	{
 		Ipponboard::FightCategory classItem;
 		m_pClassMgr->GetCategory(i, classItem);
